chore: rename highlightAutofill to highlight
