+++ .changeset/swift-pears-sleep.md
@@ -1,0 +1,9 @@
+---
+'@toptal/picasso': patch
+---
+
+---
+
+### Picasso inputs
+
+- rename `highlightAutofill` to `highlight`
