+++ packages/picasso/src/Autocomplete/Autocomplete.tsx
@@ -128,7 +128,7 @@ export interface Props
     input?: string
     disableAutofillInput?: string
   }
-  highlightAutofill?: boolean
+  highlight?: 'autofill'
 }
 
 const useStyles = makeStyles<Theme>(styles, {
@@ -178,7 +178,7 @@ export const Autocomplete = forwardRef<HTMLInputElement, Props>(
       value,
       width = 'auto',
       disabled = false,
-      highlightAutofill,
+      highlight,
       ...rest
     } = props
 
@@ -340,7 +340,7 @@ export const Autocomplete = forwardRef<HTMLInputElement, Props>(
             autoComplete={enableAutofill ? autoComplete : autoComplete || 'off'}
             testIds={testIds}
             data-testid={testIds?.input}
-            highlightAutofill={highlightAutofill}
+            highlight={highlight}
           />
         </Container>
         <div role='listbox'>
@@ -383,7 +383,6 @@ Autocomplete.defaultProps = {
   poweredByGoogle: false,
   disabled: false,
   status: 'default',
-  highlightAutofill: false,
 }
 
 Autocomplete.displayName = 'Autocomplete'
