+++ packages/picasso/src/DatePicker/DatePicker.tsx
@@ -115,7 +115,7 @@ export interface Props
   footerBackgroundColor?: string
   /** Shows orange dot indicator in days between a date range */
   indicatedIntervals?: { start: Date; end: Date }[]
-  highlightAutofill?: boolean
+  highlight?: 'autofill'
 }
 
 export const DatePicker = (props: Props) => {
@@ -147,7 +147,7 @@ export const DatePicker = (props: Props) => {
     footer,
     indicatedIntervals,
     footerBackgroundColor,
-    highlightAutofill,
+    highlight,
     ...rest
   } = props
   const classes = useStyles()
@@ -397,7 +397,7 @@ export const DatePicker = (props: Props) => {
           width={width}
           testIds={testIds}
           data-testid={testIds?.input}
-          highlightAutofill={highlightAutofill}
+          highlight={highlight}
         />
       </Container>
       {inputWrapperRef.current && (
@@ -451,7 +451,6 @@ DatePicker.defaultProps = {
   displayDateFormat: 'MMM d, yyyy',
   autoComplete: 'off',
   status: 'default',
-  highlightAutofill: false,
 }
 
 DatePicker.displayName = 'DatePicker'
