+++ packages/picasso/src/InputBase/highlightStyles.ts
@@ -1,11 +1,11 @@
 import { Theme } from '@material-ui/core/styles'
 import { alpha } from '@toptal/picasso-shared'
 
-const highlightAutofillStyles = ({ palette }: Theme) => ({
+const highlightStyles = ({ palette }: Theme) => ({
   highlightAutofill: {
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     backgroundColor: alpha(palette.yellow.lighter!, 0.6),
   },
 })
 
-export default highlightAutofillStyles
+export default highlightStyles
