+++ packages/picasso/src/RichTextEditor/styles.ts
@@ -1,7 +1,7 @@
 import { outline } from '@toptal/picasso-shared'
 import { createStyles, Theme } from '@material-ui/core/styles'
 
-import highlightAutofillStyles from '../InputBase/highlightAutofillStyles'
+import highlightAutofillStyles from '../InputBase/highlightStyles'
 
 export default (theme: Theme) => {
   const { palette, sizes } = theme
